api/serializers.py
from rest_framework import serializers
from student.models import Student

class StudentSerializer(serializers.ModelSerializer):
    class Meta:
        model = Student
        fields = ('student_name','student_matricnumber','student_course','student_age','student_gender')